@@ -5,14 +5,14 @@
 			"OCA\\QuotaWarning\\": "lib/"
 		}
 	},
-	"name": "nextcloud/notifications",
-	"description": "notifications",
+	"name": "nextcloud/quota_warning",
+	"description": "quota_warning",
 	"license": "AGPL",
 	"config": {
 		"optimize-autoloader": true,
 		"classmap-authoritative": true,
 		"platform": {
-			"php": "8.0.2"
+			"php": "8.1.31"
 		},
 		"sort-packages": true
 	},
@@ -28,9 +28,9 @@
 		"test:unit": "vendor/bin/phpunit --color -c tests/phpunit.xml"
 	},
 	"require-dev": {
-		"nextcloud/coding-standard": "^1.3",
-		"nextcloud/ocp": "dev-stable29",
-		"phpunit/phpunit": "^9.6",
-		"vimeo/psalm": "^5.26"
+		"nextcloud/coding-standard": "^1.4",
+		"nextcloud/ocp": "dev-stable31",
+		"phpunit/phpunit": "^10.5",
+		"vimeo/psalm": "^6.8"
 	}
 }
